sync from master.

+++ ci/do_circle_ci.sh
@@ -2,20 +2,24 @@
 
 set -e
 
+# Workaround for argument too long issue in protoc
+ulimit -s 16384
+
 # bazel uses jgit internally and the default circle-ci .gitconfig says to
 # convert https://github.com to ssh://git@github.com, which jgit does not support.
-mv ~/.gitconfig ~/.gitconfig_save
+if [[ -e "~/.gitconfig" ]]; then
+  mv ~/.gitconfig ~/.gitconfig_save
+fi
 
 export ENVOY_SRCDIR="$(pwd)"
 
 # xlarge resource_class.
 # See note: https://circleci.com/docs/2.0/configuration-reference/#resource_class for why we
 # hard code this (basically due to how docker works).
-export NUM_CPUS=8
+export NUM_CPUS=6
 
-# CircleCI doesn't support IPv6 by default, so we run all tests with IPv4, and
-# a limited subset with IPv6 using "machine: true" and do_circle_ci_ipv6_tests.sh
-# (see https://circleci.com/docs/2.0/executor-types/#using-machine)
+# CircleCI doesn't support IPv6 by default, so we run all tests with IPv4 only.
+# IPv6 tests are run with Azure Pipelines.
 export BAZEL_EXTRA_TEST_OPTIONS="--test_env=ENVOY_IP_TEST_VERSIONS=v4only"
 
 function finish {
@@ -27,4 +31,4 @@ trap finish EXIT
 echo "disk space at beginning of build:"
 df -h
 
-ci/do_ci.sh "$1"
+ci/do_ci.sh $*
